+++ src/com/jypec/util/MathOperations.java
@@ -13,8 +13,8 @@ public class MathOperations {
 	 * @return the maximum distance between two samples existing in a "d"-dimensional space,
 	 * bounded from (0,...,0) to (absoluteMaxValue,...,absoluteMaxValue)
 	 */
-	public static double getMaximumDistance(double absoluteMaxValue, double d) {
-		double valueIncrement = Math.sqrt(d);
+	public static float getMaximumDistance(float absoluteMaxValue, float d) {
+		float valueIncrement = (float) Math.sqrt(d);
 		return absoluteMaxValue * valueIncrement;
 	}
 
@@ -25,8 +25,8 @@ public class MathOperations {
 	 * @param base
 	 * @return the logarithm of "val" in base "base"
 	 */
-	public static double logBase(double val, double base) {
-		return Math.log10(val) / Math.log10(base);
+	public static float logBase(float val, float base) {
+		return (float) (Math.log10(val) / Math.log10(base));
 	}
